+++ app/javascript/vue/components/radials/linker/constants/links.js
@@ -192,3 +192,10 @@ export const TASK_FIELD_OCCURRENCE_MEDIA_EXTENSION = {
   label: "DwC media extension preview",
   link: "/tasks/field_occurrences/dwc_media_extension_preview",
 };
+
+export const TASK_DUPLICATE_DATA_ATTRIBUTES = {
+  label: "Duplicate DataAttributes",
+  link: "/tasks/data_attributes/duplicate_data_attributes",
+  queryParam: true,
+  saveQuery: true,
+};